Split: take away helpers from __init__.py
